requirements.txt
asgiref==3.7.2
chardet==5.2.0
cssselect2==0.7.0
Django==4.2.3
django-appconf==1.0.6
django-autoslug==1.9.9
django-bootstrap-breadcrumbs==0.9.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-debug-toolbar==4.3.0
django-filer==3.1.1
django-filter==23.2
django-js-asset==2.2.0
django-polymorphic==3.1.0
django-user-agents==0.4.0
django-views-breadcrumbs==0.1.dev3
easy-thumbnails==2.8.5
install==1.3.5
lxml==5.2.1
mysqlclient==2.2.4
Pillow==10.0.0
psycopg2==2.9.6
pytils==0.4.1
pytz==2024.1
reportlab==4.2.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
user-agents==2.2.0
webencodings==0.5.1
